requirements.txt
Flask~=2.0.0
Werkzeug~=2.0.0
SQLAlchemy~=1.4.15
WTForms~=2.3.3
Flask-Migrate~=2.7.0
Flask-Login~=0.5.0
email_validator~=1.1.2
Pillow~=8.2.0
alembic~=1.6.2